% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{aspirin_infarction_df}
\alias{aspirin_infarction_df}
\title{Aspirin after Myocardial Infarction}
\format{
A data frame with 7 observations and 6 variables:
\describe{
  \item{study}{Character variable identifying each study}
  \item{year}{Integer variable indicating the publication year}
  \item{d.asp}{Integer variable for the number of deaths in the aspirin group}
  \item{n.asp}{Integer variable for the total number of patients in the aspirin group}
  \item{d.plac}{Integer variable for the number of deaths in the placebo group}
  \item{n.plac}{Integer variable for the total number of patients in the placebo group}
}
}
\source{
Data taken from the package meta version 8.2-1
}
\usage{
data(aspirin_infarction_df)
}
\description{
This dataset, aspirin_infarction_df, is a data frame containing results from a meta-analysis
on the use of aspirin to prevent death after myocardial infarction. It includes binary outcome
data comparing aspirin and placebo groups.
}
\details{
The dataset name has been kept as 'aspirin_infarction_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
MedxR package and assists users in identifying its specific characteristics.
The suffix 'df' indicates that the dataset is a data frame. The original content has not been modified
in any way.
}
