% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{copd_drug_therapy_df}
\alias{copd_drug_therapy_df}
\title{Pharmacologic Treatments for COPD}
\format{
A data frame with 94 observations and 6 variables:
\describe{
  \item{study}{Character variable identifying each study}
  \item{year}{Integer variable indicating the publication year}
  \item{id}{Integer variable representing the study ID}
  \item{treatment}{Character variable describing the treatment group}
  \item{exac}{Integer variable indicating the number of COPD exacerbations}
  \item{total}{Integer variable for the total number of patients in the group}
}
}
\source{
Data taken from the package metadat version 1.4-0
}
\usage{
data(copd_drug_therapy_df)
}
\description{
This dataset, copd_drug_therapy_df, is a data frame containing results from 39 trials
examining pharmacologic treatments for chronic obstructive pulmonary disease (COPD).
It includes study identifiers, treatment groups, number of exacerbations, and sample sizes.
}
\details{
The dataset name has been kept as 'copd_drug_therapy_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
MedxR package and assists users in identifying its specific characteristics.
The suffix 'df' indicates that the dataset is a data frame. The original content has not been modified
in any way.
}
