% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medxr-package.R
\docType{package}
\name{MedxR}
\alias{MedxR}
\alias{MedxR-package}
\title{MedxR: Access Drug Regulatory Data via FDA and Health Canada APIs}
\description{
{
This package provides functions to access drug regulatory data from public RESTful APIs including the FDA Open API and the Health Canada Drug Product Database API, retrieving real-time or historical information on drug approvals, adverse events, recalls, and product details. Additionally, the package includes a curated collection of open datasets focused on drugs, pharmaceuticals, treatments, and clinical studies.
}
}
\details{
MedxR: Access Drug Regulatory Data via FDA and Health Canada APIs

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Access Drug Regulatory Data via FDA and Health Canada APIs.
}
\seealso{
{
Useful links:
\itemize{
  \item \url{https://github.com/lightbluetitan/medxr}
}
}
}
\author{
{
\strong{Maintainer}: Renzo Caceres Rossi \email{arenzocaceresrossi@gmail.com}
}
}
\keyword{internal}
