% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{naoh_digest_df}
\alias{naoh_digest_df}
\title{Digestibility of Straw Treated with NaOH}
\format{
A data frame with 6 observations and 3 variables:
\describe{
  \item{horse}{Integer identifier for each horse}
  \item{ordinary}{Numeric variable representing digestibility after consuming ordinary straw}
  \item{naoh}{Numeric variable representing digestibility after consuming NaOH-treated straw}
}
}
\source{
Data taken from the isdals package version 3.0.1
}
\usage{
data(naoh_digest_df)
}
\description{
This dataset, naoh_digest_df, is a data frame containing digestibility coefficients
for six horses that were fed straw. Each horse was tested under two conditions:
once after being fed ordinary straw and once after being fed straw treated with sodium hydroxide (NaOH).
}
\details{
The dataset name has been kept as 'naoh_digest_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
MedxR package and assists users in identifying its specific characteristics.
The suffix 'df' indicates that the dataset is a data frame. The original content has not been modified
in any way.
}
