% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{oral_anticoagulants_df}
\alias{oral_anticoagulants_df}
\title{Oral Anticoagulants in Coronary Artery Disease}
\format{
A data frame with 34 observations and 9 variables:
\describe{
  \item{study}{Character variable identifying each study}
  \item{year}{Integer variable indicating the publication year}
  \item{intensity}{Character variable describing treatment intensity}
  \item{asp.t}{Integer variable indicating the number of aspirin users in the treatment group}
  \item{asp.c}{Integer variable indicating the number of aspirin users in the control group}
  \item{ai}{Integer variable indicating the number of adverse events in the treatment group}
  \item{n1i}{Integer variable for the sample size of the treatment group}
  \item{ci}{Integer variable indicating the number of adverse events in the control group}
  \item{n2i}{Integer variable for the sample size of the control group}
}
}
\source{
Data taken from the package metadat version 1.4-0
}
\usage{
data(oral_anticoagulants_df)
}
\description{
This dataset, oral_anticoagulants_df, is a data frame containing results from 34 studies
evaluating the effectiveness of oral anticoagulants in patients with coronary artery disease.
The dataset includes study metadata and outcome counts for treatment and control groups, as well as intensity classifications.
}
\details{
The dataset name has been kept as 'oral_anticoagulants_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
MedxR package and assists users in identifying its specific characteristics.
The suffix 'df' indicates that the dataset is a data frame. The original content has not been modified
in any way.
}
