% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{parkinsons_list}
\alias{parkinsons_list}
\title{Dopamine Agonists in Parkinson's Disease}
\format{
A list with 5 elements:
\describe{
  \item{Outcomes}{Numeric vector representing the reduction in lost work-time for each treatment arm}
  \item{SE}{Numeric vector indicating the standard error of each observation}
  \item{Treat}{Character vector identifying the treatment type (placebo or drug)}
  \item{Study}{Numeric vector indicating the study each observation belongs to}
  \item{Treat.order}{Character vector listing the treatment names in the order of coding}
}
}
\source{
Data taken from the package bnma version 1.6.1
}
\usage{
data(parkinsons_list)
}
\description{
This dataset, parkinsons_list, is a list containing information from 7 studies investigating
the effect of dopamine agonists as adjunct therapy in patients with Parkinson's disease.
The dataset includes placebo and four active drugs coded from 2 to 5, measuring outcomes
such as lost work-time reduction.
}
\details{
The dataset name has been kept as 'parkinsons_list' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
MedxR package and assists users in identifying its specific characteristics.
The suffix 'list' indicates that the dataset is a list. The original content has not been modified
in any way.
}
