% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{pharmacy_tbl_df}
\alias{pharmacy_tbl_df}
\title{Pharmacy Client Attendance}
\format{
A tibble with 17,520 observations and 4 variables:
\describe{
  \item{date}{Date variable indicating the calendar date}
  \item{hours}{Character variable representing the hour of observation}
  \item{weekday}{Character variable representing the day of the week}
  \item{attendance}{Numeric variable indicating the number of clients observed}
}
}
\source{
Data taken from the package idarps version 0.0.5
}
\usage{
data(pharmacy_tbl_df)
}
\description{
This dataset, pharmacy_tbl_df, is a tibble containing hourly client attendance
data in a pharmacy located in Geneva, Switzerland. The dataset spans two years,
recording the number of clients per hour alongside the date and weekday.
}
\details{
The dataset name has been kept as 'pharmacy_tbl_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
MedxR package and assists users in identifying its specific characteristics.
The suffix 'tbl_df' indicates that the dataset is a tibble. The original content has not been modified
in any way.
}
